refs #621 refactor: Divide entities
