+++ assets/js/actions/projects.ts
@@ -1,30 +1,7 @@
 import { Dispatch, Action } from 'redux'
 import axios from 'axios'
-
-export type ServerProject = {
-  ID: number
-  UserID: number
-  Title: string
-  Description: string
-  RepositoryID: number | null
-  ShowIssues: boolean
-  ShowPullRequests: boolean
-}
-
-export type Project = {
-  id: number
-  userID: number
-  title: string
-  description: string
-  repositoryID: number | null
-  showIssues: boolean
-  showPullRequests: boolean
-}
-
-export type Repository = {
-  id: number
-  full_name: string
-}
+import { ServerProject, Project, converter } from '@/entities/project'
+import { Repository } from '@/entities/repository'
 
 export const RequestGetProjects = 'RequestGetProjects' as const
 export const ReceiveGetProjects = 'ReceiveGetProjects' as const
@@ -48,17 +25,7 @@ export const getProjects = () => {
   return (dispatch: Dispatch<Action>) => {
     dispatch(requestGetProjects())
     axios.get<Array<ServerProject>>('/api/projects').then(res => {
-      const data: Array<Project> = res.data.map(p => {
-        return {
-          id: p.ID,
-          userID: p.UserID,
-          title: p.Title,
-          description: p.Description,
-          repositoryID: p.RepositoryID,
-          showIssues: p.ShowIssues,
-          showPullRequests: p.ShowPullRequests
-        }
-      })
+      const data: Array<Project> = res.data.map(p => converter(p))
       dispatch(receiveGetProjects(data))
     })
   }
