+++ assets/js/actions/projects/lists/new.ts
@@ -1,6 +1,7 @@
 import axios from 'axios'
 
-import { List, ServerList, getLists } from '../lists'
+import { getLists } from '../lists'
+import { List, ServerList, converter } from '@/entities/list'
 
 export const RequestCreateList = 'RequestCreateList' as const
 export const ReceiveCreateList = 'ReceiveCreateList' as const
@@ -18,26 +19,7 @@ export const createList = (projectID: number, params: any) => {
   return async (dispatch: Function) => {
     dispatch(requestCreateList())
     return axios.post<ServerList>(`/api/projects/${projectID}/lists`, params).then(res => {
-      const data: List = {
-        id: res.data.ID,
-        user_id: res.data.UserID,
-        project_id: res.data.ProjectID,
-        title: res.data.Title,
-        color: res.data.Color,
-        list_option_id: res.data.ListOptionID,
-        is_hidden: res.data.IsHidden,
-        is_init_list: res.data.IsInitList,
-        tasks: res.data.ListTasks.map(t => ({
-          id: t.ID,
-          list_id: t.ListID,
-          user_id: t.UserID,
-          issue_number: t.IssueNumber,
-          title: t.Title,
-          description: t.Description,
-          html_url: t.HTMLURL,
-          pull_request: t.PullRequest
-        }))
-      }
+      const data = converter(res.data)
       dispatch(receiveCreateList(data))
       dispatch(getLists(projectID))
     })
