+++ assets/js/entities/repository.ts
@@ -1,0 +1,4 @@
+export type Repository = {
+  id: number
+  full_name: string
+}
